moved image files to assets folder, added link to home page in logo, some more minor fixes
