+++ src/assets/images/equipo.png
