+++ src/components/InfoAbout.js
@@ -2,17 +2,15 @@ import React from "react";
 
 function InfoAbout (props) {
     return (
-        <div className="container-img">
-          <div>
-            <img  className="help-img"
-                  src={props.img} />
-            </div>
-          <div>
-            <p>{props.titulo}</p>
-            <p>{props.texto}</p>
-          </div>
-
+      <div className="container-img">
+        <div>
+          <img className="help-img" src={props.img} alt="img" />
         </div>
+        <div>
+          <p>{props.titulo}</p>
+          <p>{props.texto}</p>
+        </div>
+      </div>
     )
 }
