+++ src/components/Nav.js
@@ -1,21 +1,23 @@
 import React from "react";
 import '../sheets-styles/Nav.css' ;
-import MainLogo from '../Imagenes/logo-principal.png' ;
+import MainLogo from "../assets/images/logo.png"
 import { Link } from "react-router-dom"
 
 function Nav() {
   return (
     <div className="main-container">
       <div className="help-container">
-        <img src={MainLogo} className="logo-principal" alt="imagen" />
+        <Link to="/">
+          <img src={MainLogo} className="logo-principal" alt="logo" />
+        </Link>
         <Link to="/about" className="help-button">
           Help Center
         </Link>
       </div>
       <div className="submit-container">
-          <Link to="/login" className="login-button">
+        <Link to="/login" className="login-button">
           Login
-          </Link>
+        </Link>
       </div>
     </div>
   )
